perf: defer auth component hydration from client:load to client:idle
HeaderAuthDesktop, HeaderAuthMobile, and MyRoastListItem all return null
until a useEffect fires to read the flag_authFeatures cookie, so their
visual output is identical whether they hydrate immediately or at idle
time. Switching from client:load to client:idle defers Clerk SDK
execution until the browser finishes initial layout, reducing main-thread
pressure on every page load.

Also extracts the duplicated useAuthFlag hook into src/hooks/useAuthFlag.ts
so the cookie-reading logic lives in one place.
