+++ src/components/header/header.astro
@@ -20,7 +20,7 @@ const isTesting = import.meta.env.PLAYWRIGHT === "true";
       </div>
 
       <div class="header-top-right">
-        {!isTesting && <HeaderAuthDesktop client:load />}
+        {!isTesting && <HeaderAuthDesktop client:idle />}
         <button
           id="theme-toggle"
           class="theme-toggle"
@@ -60,7 +60,7 @@ const isTesting = import.meta.env.PLAYWRIGHT === "true";
       <a href="/to-do-list">To-Do List</a>
       <a href="/search">Search</a>
       <a href="/archive">Archive</a>
-      {!isTesting && <HeaderAuthMobile client:load />}
+      {!isTesting && <HeaderAuthMobile client:idle />}
       <div class="dropdown">
         <a
           href="/about"
