+++ src/pages/index.astro
@@ -330,7 +330,7 @@ try {
           <img src={findARoastImg.src} alt="Find a roast" width="400" height="300" loading="lazy" />
         </a>
       </li>
-      <MyRoastListItem imgSrc={dreamingOfRoastDinners.src} client:load />
+      <MyRoastListItem imgSrc={dreamingOfRoastDinners.src} client:idle />
       <li class="heading">
         <h3><a href="/guessthescore" rel="noopener noreferrer">Guess The Score</a></h3>
         <a href="/guessthescore" rel="noopener noreferrer" class="featured-image">
